requirements.txt
flask
pycryptodome
requests
PyPDF2
reportlab